// src/shared/components/configurable-parameters/configurable-parameters.interface.ts
interface ConfigParametersId {
    id: string;
    autoHPOState: ConfigurableParametersGroupBasicParamsDTO['auto_hpo_state'];
}

type HPOOmitTypes = 'auto_hpo_state' | 'auto_hpo_value';

export interface SelectableGroupParams<T extends string | number, K extends ConfigGroupParametersDataTypesDTO>
    extends ConfigParametersId,
        Omit<SelectableGroupParamsDTO<T, K>, 'template_type' | 'data_type' | 'default_value' | HPOOmitTypes> {
    templateType: SelectableGroupParamsDTO<T, K>['template_type'];
    dataType: SelectableGroupParamsDTO<T, K>['data_type'];
    defaultValue: SelectableGroupParamsDTO<T, K>['default_value'];
    autoHPOValue: SelectableGroupParamsDTO<T, K>['auto_hpo_value'];
}

export interface NumberGroupParams
    extends ConfigParametersId,
        Omit<
            NumberGroupParamsDTO,
            'template_type' | 'data_type' | 'min_value' | 'max_value' | 'default_value' | HPOOmitTypes
        > {
    templateType: NumberGroupParamsDTO['template_type'];
    dataType: NumberGroupParamsDTO['data_type'];
    minValue: NumberGroupParamsDTO['min_value'];
    maxValue: NumberGroupParamsDTO['max_value'];
    defaultValue: NumberGroupParamsDTO['default_value'];
    autoHPOValue: NumberGroupParamsDTO['auto_hpo_value'];
}

export interface BooleanGroupParams
    extends ConfigParametersId,
        Omit<BooleanGroupParamsDTO, 'template_type' | 'data_type' | 'default_value' | HPOOmitTypes> {
    templateType: BooleanGroupParamsDTO['template_type'];
    dataType: BooleanGroupParamsDTO['data_type'];
    defaultValue: BooleanGroupParamsDTO['default_value'];
    autoHPOValue: BooleanGroupParamsDTO['auto_hpo_value'];
}

export type ConfigurableParametersParams =
    | SelectableGroupParams<string, 'string'>
    | SelectableGroupParams<number, 'integer' | 'float'>
    | BooleanGroupParams
    | NumberGroupParams;
// ...
export interface ConfigurableParametersTaskChain {
    taskId: ConfigurableParametersTaskChainDTO['task_id'];
    taskTitle: ConfigurableParametersTaskChainDTO['task_title'];
    components: ConfigurableParametersComponents[];
}

export enum ConfigurableParametersType {
    SINGLE_CONFIG_PARAMETERS = 'single-config-parameters',
    MANY_CONFIG_PARAMETERS = 'many-config-parameters',
    READ_ONLY_SINGLE_PARAMETERS = 'read-only-single-parameters',
}

export interface ConfigurableParametersSingle {
    type: ConfigurableParametersType.SINGLE_CONFIG_PARAMETERS;
    configParametersData: ConfigurableParametersTaskChain;
    updateParameter<T extends string | boolean | number>(id: string, value: T): void;
}
// ...
export interface ConfigParameterItemProp extends Partial<Pick<ConfigurableParametersSingle, 'updateParameter'>> {
    isHPO: boolean;
}